third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-number-comma-list-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <number># with iterationComposite
  assert_equals: expected "250, 275" but got "50, 75"
Harness: the test ran to completion.
